requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Django==3.0.3
django-background-tasks==1.2.5
django-compat==1.0.15
djangorestframework==3.11.0
httpie==2.0.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
Pygments==2.5.2
pylint==2.4.4
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
